Xcode Project Creation
This project compiles and initializes a window, however, it doesn't run pass that.  I made changes to accomodate for Apple's OpenGL 4.1 support, the original version of the code required 4.5 and above. Implemented functions which are missing in 4.1.  Visual Studio version of the project is now broken due to all the changes. I believe it could be easily fixed by changing the include directories, the rest of the code should run just fine.