Architecture ReWrite - First Pass
Tightning up the code having to do with Materials.   It was a bit too scattered for my taste.   From now on, Materials are in charge of assigning their own arguments to the shaders, the MaterialSettings class will be eventually removed.  The original code was re-using MaterialSettings objects to assign shader arguments no matter what shader was active, this causes assignments to shader arguments that don't exist. By encapsulating this code in each material, we avoid random assignments to shader variables that don't exist.  It also makes the code more organized, if there is ever a problem in a Material, I know exactly where to find it. 

Lastly, removing commented code from the FirstPersonController class, and finally the material store only loads shaders, the ShaderStore singleton will only expose constant methods, this is because if I ever go multithreaded, I won't have to worry about race conditions.

The work is not over yet, I thought this was a good stopping point.