FrameBuffers and Textures
In order to reduce the duplicate code in the program, I've introduced this dependency; it makes sense for frame buffers to use the texture classes to initialize their texture render targets.  I've also added a funciton to initialize multiple render targets in a framebuffer as well. 

I've also included a base texture class that can be used to put all the common functionality texture types need.