Further organizing material code
The graphics.h and graphics.cpp files acts as a manager in charge of rendering the scene.  The way I see it, some code in this class in reality should be part of the Material class, so I moved it there.  This is still work in progress and am not done yet.  I'm aware am using more texture memory in some of this code, I will do a performance/memory pass in the future to speed up the program.   For now, I'm focusing on the architecture side.  Also removing commented code.