Organizing More Code
I've moved almost all of the code that sets shader arguments to the shader classes  from the Graphics controller ( Graphics is the class that drives rendering in the application). It should be easier now to figure out where  a setting is being set on a shader by just looking at the material classes.

Removed code I don't need.

I've re-written the MaterialSetting to be a little more flexible in accepting the different value types I will be using the in applicaiton.  With MaterialGroups, am able to set any number of settings and types to any shader.  

The MeshRenderer class is now responsible for setting the materials values in a way that reflects the type of material the mesh has. We can now say a "material is beign applied to a mesh", wherareas before materials were not beign applied by the mesh itself. 

There is more code that could be organized here, but thought this was a good stopping point.  More coming in later submissions.