Cleaning up FBO construction arguments/ Creating Default FBO
This submission is mostly about reducing the number of arguments passed in to the FBO objects, I decided to make structures that group required arguments in hopefully a logical way so that the construction of these objects doesn't look so overwhelming.

Lastly, I've created a way to construct a frame buffer object that renders to the default Opengl framebuffer.  This could only be done with the 2d FBO, 3D fbos will always render off screen to a texture. 

In future submissions, I plan to use enum classes insteand of the pound defines of OpenGL to prevent for some of the constants used to set the texture state.