FBO Commands class
In the name of protecting myself from doing something dumb, I've decided to make the FBO classes more stupid proof by creating the FBO command class.  Objects from this class can only be obtained by calling FBO::Activate(), which guarantees that the FBO object is active in OpenGL before executing commands against it.  This technique could be useful for textures as well, as they also have to be activated on the OpenGL side before commands can be ran against them; may implement this for textures in the future.