3D Rendering: Step 1
I was able to render a simple box using the voxelization visualization render target.  The next step would be to try to render the voxelized scene using this same render target.  In this changelist there are minor tweaks to the FBO command class to prevent unecessarily setting the frame buffer to default.