3D Rendering: Step 3
This code will grab the rendered depth from an orthographic camera rendering and use it to fill up a 3D texture with positional data.  I also renamed a few functions.  got rid of the MeshRenderer and Renderer classes, they are no longer needed.  These two latter changes caused multiple changes around the code.  

The MaterialSetting class was renamed to ShaderParameter.h/cpp which also caused multiple changes around the code.   This name is more to the point about what this code is actually doing. 

Created a Points class which is used to push instanced points through a shader.  This code is critical to draw to 3D textrue.