Bug Fix: Depth Renderings from different perspectives did not align
When compositing the different depth buffers into the 3D voxel texture, the depth positions did not align together.  The solution was to convert from capturing camera view space to world space, and finally to the z plane camera view space, and it is in this space where the compositing happens.

Before, I was going from one camera view space to another skipping the world space transformation, this  was causing the misalignment.