Sampling 3D Texture for Cone Tracing
I've introduced new shader files where I will be implementing my version of cone tracing heavily based on the paper.  The version this code came with doesn't seem like it was following it very closely. Also wrote the base math that I'll be using to figure out the 3d textures coordinates to sample the voxel texture.