Ambient Occlusion Cone
This is the beginning to implement ambient occlusing cone tracing.  The math in this code is working as expected, all it does it takes the geometry's normal vector, creates a rotation matrix, and uses this matrix to rotate the sampling rays I feed to the shader.  More to come in later submissions.