Lighting The Voxel Texture
Before this change there was no lighting calculation, just printing of a color.  This is in preparation for the indirect illumination cone implementation. 

In the future I will be using deferred shading to store transparency and specularity.  The previous author of this shader relied on parameters passed in to use this information, but the way I've implemented voxelization has made this more difficult.