Interpolation Between LOD's
Opengl 4.1 doesn't support this feataure so I had to make the shader do it in code.  The reason I'm doing this is because previously there was too much banding happening due to the sudden switch between the different LOD samples, with this change, the banding has gone away, and both direct and indirect lighting has smoothed out.